requirements.txt
azure-functions
PyGithub=1.51
requests